refactor web-app
