+++ web-app/src/containers/Tutorial/components/Step.tsx
